[joe] Add source positions to all expressions.
Added the type ExprPos, which is an Expr parameterized with a SourcePos.
I tried to be nice and use the more generic Expr a where I could, but
many functions changed their signature to ExprPos.

In desugaring, I tried to keep the source label that is most relevant to
the expression being desugared without adding extra instances of the
same source position.  Some of these steps may need to be ironed out.

I added the sentinel value nopos that indicates no real source position
for an expr (i.e., we added it).

I added label, label', and exprLabel to extract the SourcePos from an
ExprPos.