DisableBanned works for ADsafe.get()
Note that this only works for get _as is_. If get were to be any
different, (i.e., not safe in particular ways) this checker would not be
sensitive to that difference and still let ADsafe pass.

We need to deal with stuff like this:

    if (arguments.length == 2 && !reject(object, name) {
        return object[name];
    }

    return object.prototype;

Currently, this passes the type checker. This is a general problems with
labels.