[joe] Clean up ANF.  Add VEval to the ANF data constructor.
In toANFMany, change the values inserted into the original expression to
be the *identifiers* created, not the values themselves.  This is
mainly so that large \-expressions aren't repeated.

Add the VEval constructor to the Value part of the ANF structure to
represent eval() expressions.

Appropriate changes to pretty print to accomodate VEval.

Also, always call removeHOAS before calling toANF, since toANF doesn't
expect there to be any ELet1/ELet2 expressions.