[eval] initial commit of DisableEval checker
Serveral assumptions go into this type checker. Namely:

  1. if a function's 'this' argument is 'window', no fields of 'this'
     are ever accessed.
  2. The 'Function' identifier (function constructor) and the 'eval'
     identifier are unsafe, and all other global identifiers are safe to
     apply.
  3. All global identifiers, if treated as object, have unsafe fields:
       eval, constructor, $proto, and prototype.

For a value to be unsafe means that if it is applied, it may result in a
call to eval.

The first assumption is fairly clear by inspecting the code of ADsafe.
However, it should be verified through some means.

The first part of the second assumption is obvious from inspecting the
ECMA environment. The second part is not at all clear, though likely.

The third assumption may underapproximate.

All of these issues, and likely more, would not be an issue if the type
system ran on ADsafe together with the global environment. Currently it
does not, as it has issues with the following:

  1. $makeException's first argument is looked up in $global. Though the
     result (an exception constructor) is always safe, there is no way
     to know that since typing information from an application is not used
     when typing the body of a function.

     A solution to this problem is to have $makeException take the
     constructor itself rather than a string. Naïvely changin the code
     to do this will not work, as the constructors are not in the scope
     of $makeExcpetion. Additionally, $makeException is not in scope
     within the body of $makeException, though it is referenced in the
     body.

  2. Identifers are declared as references to undefined or empty objects,
     and then then the objects in the reference are incrementally
     updated with new fields. Those fields are sometimes unsafe, which
     causes the checker to fail on the assignments.