[joe] Use occurrence typing to simplify redundant ifs, and an ANF fix.
In ANF, don't make a bunch of unneccesary ALet forms around BOp,
BObject, etc.  Rewrite of toANFMany to accomplish this.  This was
necessary to get the occurrence typing to work -- needed the actual
relevant variable to be in the typeof expression.

SimplifyIf.hs (possibly a better name for this file needed?) simplifies
expressions of the form:

if((typeof x) === "number") {
	if((typeof x) === "string") {
	[dead code]
	}
        else {
	[live code]
	}
}

to

if((typeof x) === "number") {
	[live code in a thunk]
}

The checking is more general than this, but this is the application of
it that is in this commit.  The thunk is necessary because of the
structure of the ANF syntax.