[adsafe] Fixes and testing for ANF.
Changes to ANF:

-Removed the data constructor ARec, and changed transformation of EWhile
to use ALet instead.

-Added the data constructor ASeq, and used it in the transformation of
ESeq.

-Fixes to AFinally and ACatch

Changes to Disable*, PrettyPrint, and SimplifyIf reflect these updates.
ASeq uses |begin| just like ESeq for pretty printing to interpreted
code.

In the interpreter, we install a handler at the top of each procedure
that catches |break| statements inside the body of the procedure.  We
only want it to do this for Javascript functions, not for procedures we
introduce (like with desugared |while|).  So, a special check is added
on the arity of the function -- if the procedure expects and receives no
arguments (e.g. it was introduced in desugaring or a later
transformation), don't add the handler.  Without this, there are
exceptions of the form "break crossing a procedure boundary".

impl.lambdajs.anf is a shell script for running the ANF executable
(in the ADsafe directory).  Useful for adding in spidermonkey_test.sh.