[\JS] Eliminate redundant let-bindings in ANF
When ANFing a let expression, sometime the right-hand side of a clause
is already a BindExp and does not need to ANFed. In particular, there is
no need to introduce an intermediate binding for the result of that
expression.

Currently we only handle a few common cases, such as alloc, deref, and
object literals. What really should happen is that ANFing should be
refactored so that the continuation takes either a Value or a BindExp.
let (and letrec if we bring it back) could make use of the BindExp to
reduce code, and the all other forms can introduce a let to name the
expression and use the generated identifer as a value.