[adsafe] New checker for Window escaping.
Implementation of a type checker for a different (better) type system
for guarding against window escapes.

The type hierarchy looks like:

             --This
            /
    --Window---WinOf x
   /
  /       --AString s (constant strings)
 /       /
JS---Safe---IsWin x [T/F]
  \
   --MaybeWin x [T/F]

Supertypes are further to the left in the hierarchy.

We're still interested in the expression

(this === this.window)

Which is made up of an expression of type [This] and an expression of
type [WinOf "this"].

The checker does if-splitting on this expression.  In the normal form,
it indicates that "this" refers to the global object in the
then-branch, and is Safe in the else-branch.

Some other checks are also done, that involve negations or
combinations of this check with others.  MaybeWin accounts for typing
thsese cases:

(this === this.window || e) : MaybeWin "this" True

(this !== this.window && e) : MaybeWin "this" False

Now, instead of assuring that nothing of type Window escapes, the
checker assures that everything that escapes is a subtype of Safe.  It
runs successfully on ADSafe.