[adsafe] Add the PushDown module.
Sort of the opposite of let-lifting.  PushDown takes lets of one
identifier to another and eliminates them by using the earlier
identifier in all places where the other was free in the body of the
let.  For example:

(begin
  (set! x 4)
  (let (y x)
     (throw y))))

=>

(begin
   (set! x 4)
   (throw x)))

This is really helpful in cases where temporary identifiers shadow the
identifiers in real code that we want to accumulate type information
about.  In particular, I'm finding this important for the MakeableTags
whitelist checker.