[adsafe] MakeableTags whitelist checker
Assumptions:

- "createElement" is the only field of document that we need to check
  for.  There are likely others, but ADSafe wraps calls to *this* one,
  so that's what the checker in the first version checks.

- (deref $global "document") is the only way the document object is
  accessed.

- makeableTagName is a location that holds an object.

This also requires a number of pre-checking transformations.  I have
been using:

pushDown . removeUnused . ifReduce . removeUnused . ifReduce . anf

And the checker succeeds on ADSafe.  Succeeding means that every time
document.createElement is called, the (non-this) arguments all have
type "Makeable".

The type "Makeable" on an identifier x indicates that the value of x
is a string, and makeableTagName[x] === true.