[analysis] MergeSequences module and related export.
In ANFing, the ESeq construct is converted into an ASeq one, except in
special cases where the first expression is immediately a value or a
bind expression.  This is to accommodate non-local control operators,
like break and throw.

This leads to many cases where the e1 in (begin e1 e2) is turned into
a series of nested lets with a value or bind expression as the body.
However, having it in the ASeq form makes type analysis difficult.
This module takes sequences where the first expression fits this
pattern, and wraps e2 in the nested lets of e1, with an additional let
(to an unused variable) for the innermost expression of e1.

This facilitates better type analysis, and makes let-lifting possible
in certain cases.