[anf] ANF takes a parameter to use as its variable prefix.
Since often it's useful to ANF with a bunch of optimizations once, and
then parse and ANF again to go through a checker, give those two as
options.

When doing this, it's nice to have the different stages of ANFing have
different variable prefixes.  Allow this to happen, with related
changes to the main functions to have sensible defaults.