[anf] Brought anf much closer to idempotence.
I decided to do this pretty naively, but it works well.  Specifically,
just wrote predicates and transformers for Exprs that correspond
directly to BindExps and Values.

This is still missing some cases that I can't pin down, but only ~90
extra variables are added over all of ANF'd ADsafe.

Passes the usual test cases.