[anf] Undo the changes from 0f06f067, replace with better fixes.
There were two main problems with ANF'ing.

1:

(let ((x (let (shared 5) shared))
      (y (let (shared 10) shared)))
  (+ x y))

-->

(let (shared 5)
  (let (shared 10)
    (let ((x shared)
          (y shared))
      (+ x y))))

This is fixed by doing alpha renaming.

This could also be fixed by unpacking the let list:

(let (shared 5)
  (let (x shared)
    (let (shared 10)
      (let (y shared)
         (+ x y)))))

2:

The second expression in an ESeq should be toANFValue'd, not just
toANF'd -- sometimes this resulted in a BindExp not showing up.  This
is why the interpreter was running out of memory -- the set! at the
end of a while loop wasn't happening, so the evaluation context would
grow unbounded.

Also changed all the BindExp cases of toANF to return

(k $ Right [[bind-exp]])

This matches what was originally intended by toValue -- all lets are
introduced here, when BindExps need to be named to use in another
expression.